cart.py и views.py переделаны под int

+++ apps/shop/cart.py
@@ -1,4 +1,3 @@
-from decimal import Decimal
 from django.conf import settings
 from .models import Product
 
@@ -51,7 +50,7 @@ class Cart(object):
             self.cart[str(product.id)]['product'] = product
 
         for item in self.cart.values():
-            item['price'] = Decimal(item['price'])
+            item['price'] = int(item['price'])
             item['total_price'] = item['price'] * item['quantity']
             yield item
 
@@ -65,7 +64,7 @@ class Cart(object):
         """
         Подсчет стоимости товаров в корзине.
         """
-        return sum(Decimal(item['price']) * item['quantity'] for item in self.cart.values())
+        return sum(int(item['price']) * item['quantity'] for item in self.cart.values())
 
     def clear(self):
         # удаление корзины из сессии
